合并 LaunchMain 与 Launch
