+++ src/main/java/snw/kookbc/impl/mixin/Blackboard.java
@@ -27,7 +27,7 @@ package snw.kookbc.impl.mixin;
 import org.spongepowered.asm.service.IGlobalPropertyService;
 import org.spongepowered.asm.service.IPropertyKey;
 
-import static snw.kookbc.impl.launch.Launch.blackboard;
+import static snw.kookbc.LaunchMain.blackboard;
 
 @SuppressWarnings("unchecked")
 public class Blackboard implements IGlobalPropertyService {
