+++ src/main/java/snw/kookbc/impl/plugin/PluginMixinConfigManager.java
@@ -19,7 +19,7 @@ package snw.kookbc.impl.plugin;
 
 import org.spongepowered.asm.mixin.Mixins;
 import snw.jkook.plugin.PluginDescription;
-import snw.kookbc.impl.launch.Launch;
+import snw.kookbc.LaunchMain;
 
 import java.io.File;
 import java.io.IOException;
@@ -36,7 +36,7 @@ public class PluginMixinConfigManager {
     public PluginMixinConfigManager() throws IOException {
         this.tempDir = Files.createTempDirectory("KookBC-Mixin").toFile();
         this.tempDir.deleteOnExit();
-        Launch.classLoader.addURL(tempDir.toURI().toURL());
+        LaunchMain.classLoader.addURL(tempDir.toURI().toURL());
     }
 
     public void add(PluginDescription description, String name, InputStream stream) throws IOException {
